notes/app/requirements.txt
fastapi
sqlalchemy
python-jose
passlib==1.7.4
bcrypt==4.0.1
pydantic
psycopg2-binary
kafka-python